refactor(xtask): isolate test target policy
Keep Cargo metadata parsing and its regression coverage in a focused module.
